move addTodoSuccess and addTodo to actions/todo.js
